+++ src/index.ts
@@ -1,8 +1,7 @@
 import "./style/theme.css";
 import "./style/core.css";
-import "./style/classicVariant.css";
-import "./style/defaultVariant.css";
-import "./style/minimalVariant.css";
+import "./style/icon.css";
+import "./style/variant.css";
 
 export { showNotify } from "./core/showNotify";
 export { setNotifyDefaults } from "./core/notifyGlobals";